add tests for the begin and end options. Close #10.
