+++ Cargo.toml
@@ -29,3 +29,4 @@ lazy_static = "1.4.0"
 structopt = "0.3.21"
 shellexpand = "2.1.0"
 two_timer = "2.2.0"
+assert_cmd = "1.0.3"
